+++ src/CocoTBFramework/components/wb4/__init__.py
@@ -1,0 +1,1 @@
+"""Wishbone B4 (pipelined) BFMs: packet, master, slave, monitor, factories."""